long policy
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@5947 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/include/stdlib.h
@@ -32,8 +32,8 @@ void *realloc(void *, size_t);
 
 #define RAND_MAX 0x7fffffff
 
-void srand(unsigned int seed);
-int rand(void);
+void srand(unsigned long seed);
+long rand(void);
 
 #define abs(x) ((x)>0?(x):-(x))
